Added read tests; looks ready to go
Updated README.md

Started to use C99 in the test suite; I was missing vsnprintf

Added a cmp_mp_version function which returns the version of the
MessagePack specification currently in use.  As there's no 100% way to
tell what the version of the spec is on the MP site; this is pretty
useless, and I might even remove it.

Removed some test cases that were just wrong (int overflows, etc.)

Made some small changes so GCC will compile the test suite without
errors.