Make byte-wise float comparison intentional.
I'm less sure about this PR than the others. On the one hand, this makes the intention of
byte-wise float compare more obvious (and avoids a compiler warning `-Wfloat-equal`).
On the other hand, it adds a dependency on the `memcmp` function and the `string.h`
header, which hasn't been needed so far, and it's quite a nice property to only be depending
on very simple type headers and no C stdlib functions at all. Also, `memcmp` has less
type-safety.

So: feel free to reject this if you agree more strongly with the counter points.