Correct `is_bigendian()` to do as the readme says.
This is a semantic change: previously, if `WORDS_BIGENDIAN` was not defined, the code would
statically default to little endian by assuming `WORDS_BIGENDIAN == 0`. With this change, if
`WORDS_BIGENDIAN` is not defined, the code will default to runtime detection, adding a memory
lookup (hopefully cached) and a branch (hopefully branch-predicted correctly).

This is the documented behaviour in README.md. To avoid performance regressions, users of the
library should always explicitly compile with `-DWORDS_BIGENDIAN=0`.